pyproject.toml
[project]
name = "asar-seastate-processor"
requires-python = ">= 3.11"
description = "Python library to generate Level-2 WAVE Product starting from Ifremer Level-1B or Level-1C ENVISAT ASAR files"
readme.file = "README.md"
readme.content-type = "text/markdown"
license = {text = "MIT"}
keywords = ["SAR", "wave", "research", "sea-state"]
authors = [
  {name = "Robin Marquart"},
  {name = "Antoine Grouazel"},
  {name = "Quentin Febvre"}
]
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = [
    "xarray",
    "h5netcdf",
    "netCDF4",
    "pyyaml",
    "onnxruntime"
]

dynamic = ["version"]

[build-system]
requires = ["setuptools>=64.0", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["asar_seastate_processor"]

[tool.setuptools.package-data]
"*" = ["config/*.yaml", "models/*.onnx"]

[tool.setuptools_scm]
fallback_version = "999"

[tool.isort]
profile = "black" 
skip_gitignore = true
float_to_top = true
default_section = "THIRDPARTY"
known_first_party = "asar-seastate-processor"

[project.scripts]
ASAR-L2-wave-processor = "asar_seastate_processor.main:main"